requirements.txt
biopython==1.78
comtypes==1.1.7
h5py==3.1.0
lxml==4.6.2
matplotlib==3.3.3
networkx==2.5
numba==0.52.0
numpy==1.19.4
pandas==1.2.1
psutil==5.8.0
pyyaml==5.3.1
scikit-learn==0.23.2
scipy==1.6.0
seaborn==0.11.1
tables==3.6.1
tqdm==4.56.0
SQLAlchemy==1.3.22
pythonnet==2.5.1
fastcore==1.3.5
nbdev==1.1.4
click==7.1.2
streamlit==0.80.0
plotly_express==0.4.0
watchdog==2.1.0
wget==3.2
